Version 03 - Complete
Changed the folder structure that the script requires (compared to the last 2 versions). Significant improvements to the image and ROI handling to iterate through all the images without constantly closing and opening again. The quantification (Measure for area, and Analyze Particles) now is optional and includes the option to do it with Find Maxima too. All the parameters needed for re-analysis are asked to the user through the menu and the result tables were improved to merge all the parameters with their labels (image and ROI name) into a single csv file. For easy preview of the results in case re-analysis with different parameters is needed, the script saves the colour-coded particle counts of the Analyze Particles plug-in (any black particles are not counted due to the initial parameters).

Pending: Enclose code into a function to enable batch mode (multiple folders). Follow-up script using python-pptx to import all the cropped images produced by this script for easier result presentation.