Version 01 - Complete
Notebook 03 is the same as the one in Tools for students/Eduardo/ with the same name, just edited the directories to make it work from the Mulligan lab Google Drive account. Notebook 04 is a similar tool, but with the most updated dataset (it includes SY5Y cells!) and also has directories referenced to the Mulligan Google account.