Version 04 - Complete
Added short parameter that needs to be passed on to the Analyze Particles plug-in for images that are calibrated in inches. Added option to pass a set of particle size and circularity different for each method. Also, there is now a Colab notebook to generate a pptx from the outputs of this script.